requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.88.0
requests>=2.28.0
plotly>=5.17.0
Pillow>=9.0.0